+++ core/connectors/notion/client.js
@@ -15,3 +15,4 @@ function createNotionHttpClient({ notionToken }) {
 module.exports = { createNotionHttpClient };
 
 
+