+++ core/connectors/notion/ideas_repo.js
@@ -166,3 +166,4 @@ class NotionIdeasRepo {
 module.exports = { NotionIdeasRepo };
 
 
+