+++ core/connectors/postgres/reminders_repo.js
@@ -56,3 +56,4 @@ class RemindersRepo {
 module.exports = { RemindersRepo };
 
 
+