+++ core/connectors/stt/openai_whisper.js
@@ -31,3 +31,4 @@ async function transcribeWavWithOpenAI({ apiKey, wavPath, model = 'whisper-1', l
 module.exports = { transcribeWavWithOpenAI };
 
 
+