+++ core/connectors/telegram/files.js
@@ -30,3 +30,4 @@ async function downloadTelegramFileToTmp({ bot, fileId, prefix = 'tg_voice', ext
 module.exports = { downloadTelegramFileToTmp };
 
 
+